tlsscan: respect timeout if TLS handshake hangs
Up until this commit the context supplied during the scan only applied
to the initial dial of the connection. It is my suspicion that there are
some servers which will not close the connection if given something they
do not expect (a TLS handshake).

We should be good citizens anyway and not depend on servers having that
behavior in the first place.

[fixes #153764270]