tlsscan: listener hijinks are no longer required
Use a standard httptest.Server as a fixture for the mutual TLS test.
Whatever problem was causing issues in the past with our mutual TLS
configuration has been solved. The creation of an unmonitored child
goroutine was causing flakes in CI due to us not knowing when it was
ready to serve requests.

Signed-off-by: Christopher Brown <cbrown@pivotal.io>